evito di cancellare il DUMMYFILE dentro collectionData all'avvio del server + fix gestione indici JSOnComposer

+++ SWE/package_npm/maap_server/modelServer/DSL/DSLManager.js
@@ -47,7 +47,8 @@ var deleteFolderRecursive = function(path) {
       if(fs.lstatSync(curPath).isDirectory()) { // recurse
         deleteFolderRecursive(curPath);
       } else { // delete file
-        fs.unlinkSync(curPath);
+        if (curPath.indexOf('DUMMYFILE') == -1) 
+			fs.unlinkSync(curPath);
       }
     });
 	//evito di eliminare la directory
